+++ test/dummy/app/models/dummy/session_attachment/document.rb
@@ -1,0 +1,4 @@
+# frozen_string_literal: true
+
+class Dummy::SessionAttachment::Document < Folio::SessionAttachment::Document
+end