+++ test/dummy/app/models/dummy/session_attachment/image.rb
@@ -1,0 +1,4 @@
+# frozen_string_literal: true
+
+class Dummy::SessionAttachment::Image < Folio::SessionAttachment::Image
+end